Added more request checks and updated user related actions

+++ beFit/api/1.presentation/src/main/java/com/proiect/controllers/abonament/AbonamentController.java
@@ -1,6 +1,9 @@
 package com.proiect.controllers.abonament;
 
 import com.proiect.entities.Abonament;
+import com.proiect.entities.Role;
+import com.proiect.exceptions.UserOperationNotAllowedException;
+import com.proiect.security.JwtTokenProvider;
 import com.proiect.services.abonament.IAbonamentService;
 import com.proiect.services.models.abonament.AbonamentModel;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,15 +21,21 @@ public class AbonamentController {
 
     @Autowired
     private final IAbonamentService abonamentService;
+    @Autowired
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public AbonamentController(IAbonamentService abonamentService) {
+    public AbonamentController(IAbonamentService abonamentService, JwtTokenProvider jwtTokenProvider) {
         this.abonamentService = abonamentService;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @GetMapping("/")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Abonament>> listAll() {
-        return new ResponseEntity<>(abonamentService.listAll(), HttpStatus.OK);
+    public ResponseEntity<List<Abonament>> listAll(@RequestHeader(value = "Authorization") String Authorization) {
+        var role = jwtTokenProvider.getRole(Authorization.substring(7)).get(0);
+        if (role.get("authority").equals(Role.ROLE_ADMIN.getAuthority())) {
+            return new ResponseEntity<>(abonamentService.listAll(), HttpStatus.OK);
+        } else throw new UserOperationNotAllowedException("You can't do that.");
     }
 
     @GetMapping("/{id}")
@@ -37,21 +46,33 @@ public class AbonamentController {
 
     @PostMapping("/")
     @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<Abonament> insert(@Validated @RequestBody AbonamentModel abonamentModel){
-        return new ResponseEntity<>(abonamentService.insert(abonamentModel), HttpStatus.CREATED);
+    public ResponseEntity<Abonament> insert(@Validated @RequestBody AbonamentModel abonamentModel,
+                                            @RequestHeader(value = "Authorization") String Authorization) {
+        var role = jwtTokenProvider.getRole(Authorization.substring(7)).get(0);
+        if (role.get("authority").equals(Role.ROLE_ADMIN.getAuthority())) {
+            return new ResponseEntity<>(abonamentService.insert(abonamentModel), HttpStatus.CREATED);
+        } else throw new UserOperationNotAllowedException("You can't do that.");
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.ACCEPTED)
-    public ResponseEntity<Abonament> update(@PathVariable int id, @Validated @RequestBody
-            AbonamentModel abonamentModel) {
+    public ResponseEntity<Abonament> update(@PathVariable int id,
+                                            @Validated @RequestBody AbonamentModel abonamentModel,
+                                            @RequestHeader(value = "Authorization") String Authorization) {
+        var role = jwtTokenProvider.getRole(Authorization.substring(7)).get(0);
+        if (role.get("authority").equals(Role.ROLE_ADMIN.getAuthority())) {
         return new ResponseEntity<>(abonamentService.update(id, abonamentModel), HttpStatus.ACCEPTED);
+        } else throw new UserOperationNotAllowedException("You can't do that.");
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable int id,
+                       @RequestHeader(value = "Authorization") String Authorization) {
+        var role = jwtTokenProvider.getRole(Authorization.substring(7)).get(0);
+        if (role.get("authority").equals(Role.ROLE_ADMIN.getAuthority())) {
         abonamentService.delete(id);
+        } else throw new UserOperationNotAllowedException("You can't do that.");
     }
 
 }
